vendor-esr: Import FIREFOX_38_1_0esr_RELEASE from vendor.
Source URL: http://hg.mozilla.org/releases/mozilla-esr38/archive/61f7702f2ddc.tar.bz2

+++ media/webrtc/trunk/webrtc/modules/audio_processing/aec/include/echo_cancellation.h
@@ -14,32 +14,32 @@
 #include "webrtc/typedefs.h"
 
 // Errors
-#define AEC_UNSPECIFIED_ERROR           12000
-#define AEC_UNSUPPORTED_FUNCTION_ERROR  12001
-#define AEC_UNINITIALIZED_ERROR         12002
-#define AEC_NULL_POINTER_ERROR          12003
-#define AEC_BAD_PARAMETER_ERROR         12004
+#define AEC_UNSPECIFIED_ERROR 12000
+#define AEC_UNSUPPORTED_FUNCTION_ERROR 12001
+#define AEC_UNINITIALIZED_ERROR 12002
+#define AEC_NULL_POINTER_ERROR 12003
+#define AEC_BAD_PARAMETER_ERROR 12004
 
 // Warnings
-#define AEC_BAD_PARAMETER_WARNING       12050
+#define AEC_BAD_PARAMETER_WARNING 12050
 
 enum {
-    kAecNlpConservative = 0,
-    kAecNlpModerate,
-    kAecNlpAggressive
+  kAecNlpConservative = 0,
+  kAecNlpModerate,
+  kAecNlpAggressive
 };
 
 enum {
-    kAecFalse = 0,
-    kAecTrue
+  kAecFalse = 0,
+  kAecTrue
 };
 
 typedef struct {
-    int16_t nlpMode;        // default kAecNlpModerate
-    int16_t skewMode;       // default kAecFalse
-    int16_t metricsMode;    // default kAecFalse
-    int delay_logging;      // default kAecFalse
-    //float realSkew;
+  int16_t nlpMode;      // default kAecNlpModerate
+  int16_t skewMode;     // default kAecFalse
+  int16_t metricsMode;  // default kAecFalse
+  int delay_logging;    // default kAecFalse
+  // float realSkew;
 } AecConfig;
 
 typedef struct {
@@ -50,10 +50,10 @@ typedef struct {
 } AecLevel;
 
 typedef struct {
-    AecLevel rerl;
-    AecLevel erl;
-    AecLevel erle;
-    AecLevel aNlp;
+  AecLevel rerl;
+  AecLevel erl;
+  AecLevel erle;
+  AecLevel aNlp;
 } AecMetrics;
 
 struct AecCore;
@@ -68,7 +68,7 @@ extern "C" {
  *
  * Inputs                       Description
  * -------------------------------------------------------------------
- * void **aecInst               Pointer to the AEC instance to be created
+ * void**  aecInst              Pointer to the AEC instance to be created
  *                              and initialized
  *
  * Outputs                      Description
@@ -76,28 +76,28 @@ extern "C" {
  * int32_t return               0: OK
  *                             -1: error
  */
-int32_t WebRtcAec_Create(void **aecInst);
+int32_t WebRtcAec_Create(void** aecInst);
 
 /*
  * This function releases the memory allocated by WebRtcAec_Create().
  *
  * Inputs                       Description
  * -------------------------------------------------------------------
- * void         *aecInst        Pointer to the AEC instance
+ * void*        aecInst         Pointer to the AEC instance
  *
  * Outputs                      Description
  * -------------------------------------------------------------------
  * int32_t      return          0: OK
  *                             -1: error
  */
-int32_t WebRtcAec_Free(void *aecInst);
+int32_t WebRtcAec_Free(void* aecInst);
 
 /*
  * Initializes an AEC instance.
  *
  * Inputs                       Description
  * -------------------------------------------------------------------
- * void           *aecInst      Pointer to the AEC instance
+ * void*          aecInst       Pointer to the AEC instance
  * int32_t        sampFreq      Sampling frequency of data
  * int32_t        scSampFreq    Soundcard sampling frequency
  *
@@ -106,15 +106,15 @@ int32_t WebRtcAec_Free(void *aecInst);
  * int32_t        return        0: OK
  *                             -1: error
  */
-int32_t WebRtcAec_Init(void *aecInst, int32_t sampFreq, int32_t scSampFreq);
+int32_t WebRtcAec_Init(void* aecInst, int32_t sampFreq, int32_t scSampFreq);
 
 /*
  * Inserts an 80 or 160 sample block of data into the farend buffer.
  *
  * Inputs                       Description
  * -------------------------------------------------------------------
- * void           *aecInst      Pointer to the AEC instance
- * int16_t        *farend       In buffer containing one frame of
+ * void*          aecInst       Pointer to the AEC instance
+ * const float*   farend        In buffer containing one frame of
  *                              farend signal for L band
  * int16_t        nrOfSamples   Number of samples in farend buffer
  *
@@ -123,8 +123,8 @@ int32_t WebRtcAec_Init(void *aecInst, int32_t sampFreq, int32_t scSampFreq);
  * int32_t        return        0: OK
  *                             -1: error
  */
-int32_t WebRtcAec_BufferFarend(void *aecInst,
-                               const int16_t *farend,
+int32_t WebRtcAec_BufferFarend(void* aecInst,
+                               const float* farend,
                                int16_t nrOfSamples);
 
 /*
@@ -132,10 +132,10 @@ int32_t WebRtcAec_BufferFarend(void *aecInst,
  *
  * Inputs                       Description
  * -------------------------------------------------------------------
- * void          *aecInst       Pointer to the AEC instance
- * int16_t       *nearend       In buffer containing one frame of
+ * void*         aecInst        Pointer to the AEC instance
+ * float*        nearend        In buffer containing one frame of
  *                              nearend+echo signal for L band
- * int16_t       *nearendH      In buffer containing one frame of
+ * float*        nearendH       In buffer containing one frame of
  *                              nearend+echo signal for H band
  * int16_t       nrOfSamples    Number of samples in nearend buffer
  * int16_t       msInSndCardBuf Delay estimate for sound card and
@@ -146,18 +146,18 @@ int32_t WebRtcAec_BufferFarend(void *aecInst,
  *
  * Outputs                      Description
  * -------------------------------------------------------------------
- * int16_t       *out           Out buffer, one frame of processed nearend
+ * float*        out            Out buffer, one frame of processed nearend
  *                              for L band
- * int16_t       *outH          Out buffer, one frame of processed nearend
+ * float*        outH           Out buffer, one frame of processed nearend
  *                              for H band
  * int32_t       return         0: OK
  *                             -1: error
  */
-int32_t WebRtcAec_Process(void *aecInst,
-                          const int16_t *nearend,
-                          const int16_t *nearendH,
-                          int16_t *out,
-                          int16_t *outH,
+int32_t WebRtcAec_Process(void* aecInst,
+                          const float* nearend,
+                          const float* nearendH,
+                          float* out,
+                          float* outH,
                           int16_t nrOfSamples,
                           int16_t msInSndCardBuf,
                           int32_t skew);
@@ -167,7 +167,7 @@ int32_t WebRtcAec_Process(void *aecInst,
  *
  * Inputs                       Description
  * -------------------------------------------------------------------
- * void           *handle       Pointer to the AEC instance
+ * void*          handle        Pointer to the AEC instance
  * AecConfig      config        Config instance that contains all
  *                              properties to be set
  *
@@ -183,11 +183,11 @@ int WebRtcAec_set_config(void* handle, AecConfig config);
  *
  * Inputs                       Description
  * -------------------------------------------------------------------
- * void           *handle       Pointer to the AEC instance
+ * void*          handle        Pointer to the AEC instance
  *
  * Outputs                      Description
  * -------------------------------------------------------------------
- * int            *status       0: Almost certainly nearend single-talk
+ * int*           status        0: Almost certainly nearend single-talk
  *                              1: Might not be neared single-talk
  * int            return         0: OK
  *                              -1: error
@@ -199,11 +199,11 @@ int WebRtcAec_get_echo_status(void* handle, int* status);
  *
  * Inputs                       Description
  * -------------------------------------------------------------------
- * void           *handle       Pointer to the AEC instance
+ * void*          handle        Pointer to the AEC instance
  *
  * Outputs                      Description
  * -------------------------------------------------------------------
- * AecMetrics     *metrics      Struct which will be filled out with the
+ * AecMetrics*    metrics       Struct which will be filled out with the
  *                              current echo metrics.
  * int            return         0: OK
  *                              -1: error
@@ -232,13 +232,13 @@ int WebRtcAec_GetDelayMetrics(void* handle, int* median, int* std);
  *
  * Inputs                       Description
  * -------------------------------------------------------------------
- * void           *aecInst      Pointer to the AEC instance
+ * void*          aecInst       Pointer to the AEC instance
  *
  * Outputs                      Description
  * -------------------------------------------------------------------
  * int32_t        return        11000-11100: error code
  */
-int32_t WebRtcAec_get_error_code(void *aecInst);
+int32_t WebRtcAec_get_error_code(void* aecInst);
 
 // Returns a pointer to the low level AEC handle.
 //
