+++ xpcom/glue/nsTArrayForwardDeclare.h
@@ -1,5 +1,5 @@
-/* -*- Mode: C++; tab-width: 2; indent-tabs-mode: nil; c-basic-offset: 2 -*- */
-/* vim:set ts=2 sw=2 sts=2 et cindent: */
+/* -*- Mode: C++; tab-width: 8; indent-tabs-mode: nil; c-basic-offset: 2 -*- */
+/* vim: set ts=8 sts=2 et sw=2 tw=80: */
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
@@ -19,7 +19,7 @@
 // template parameters.
 //
 
-#include <stdint.h>
+#include <stddef.h>
 
 template<class E>
 class nsTArray;
@@ -27,10 +27,10 @@ class nsTArray;
 template<class E>
 class FallibleTArray;
 
-template<class E, uint32_t N>
+template<class E, size_t N>
 class nsAutoTArray;
 
-template<class E, uint32_t N>
+template<class E, size_t N>
 class AutoFallibleTArray;
 
 #define InfallibleTArray nsTArray
